- fixup ListView WinUI
