+++ src/Compatibility/Core/src/Handlers/ListView/Windows/ICellRenderer.cs
