+++ src/Compatibility/Core/src/Handlers/ListView/Windows/TextCellRenderer.cs
